Move code specifically for local running to example_project
